Autocomplete: Remove multilineMode in favor of a simple boolean (#329)
Renaming all occurrences of `multilineMode` to a `multiline` boolean
flag instead. My idea of having more modes in the future didn't pan out
and this was just adding confusion.

The old field name is still kept in analytics for a bit so we can keep
comparing data with the past. We can rm that in the future once we have
more data with the new format.

## Test plan

Tests are green

<!--
All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Some examples:

// Just a doc change
none - docs change

// Unit tests got your back?
Unit tests

// Tested it manually and CI will also pitch in?
Manually tested and CI
-->