support working with recipes, etc., on virtual files
Recipes that provide information about code (such as an explanation) should work on read-only files.

Recipes that modify files in-place should use the VS Code `vscode.workspace.fs.isWritableFileSystem` extension API to see if the document's URI scheme is writable. This commit does NOT update all recipes to check that.