Autocomplete: Cache previously cancelled completions (#317)
Closes #298
Closes #315
Closes #292

This PR adds a new `RequestManager`. This class is doing the bookkeeping
of all completion network requests.

The idea here is that we no longer cancel network requests after they
are triggered. Instead, we complete the requests even when the prefix
was updated already and use the result in the completions cache. The
cache can then synthesise potential completions for requests in the
future.

An example is something like this. Imagine we are hitting the debounce
timeout with this prefix:

```
console.
```

Now we start a network request for the completion but the user decides
to continue typing before the response arrives and changes the prefix to
something like this:

```
console.log(
```

Now, briefly after the second pause, the firsts request comes back with
the completion of `log("Hello world")` and we put it into the
completions cache. We then recheck all open requests, including the one
with the `console.log(` suffix, against the cache again which will cause
a hit. We can now respond with the second completion even though the
network request is not finished yet.

## Test plan

- Added tests

<!--
All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Some examples:

// Just a doc change
none - docs change

// Unit tests got your back?
Unit tests

// Tested it manually and CI will also pitch in?
Manually tested and CI
-->