simplify EventLogger.log to only take public event props
Previously, the event logger took 2 args: event properties and public properties. The former was a superset of the latter and contained private information that could end up in, for example, the audit log. This is not needed for Cody's telemetry. See https://sourcegraph.slack.com/archives/CN4FC7XT4/p1690356187459289 for more context.

Also removes unnecessary passing along of the serverEndpoint (which is added globally).