refactor telemetry service for VS Code extension (#388)
There is no significant change to the telemetry event data that is sent
up (except for a few reductions in the event properties when we were
sending up complex objects).

- Add a few more tracked events related to the signup flow.
- Introduce a new `TelemetryService` with the same API as `logEvent` but
that is not global.
- Use the `TelemetryService` interface in the VS Code webviews as well
instead of manually passing events.
- Clean up other code.
- Simplify EventLogger.log to only take public event props. Previously,
the event logger took 2 args: event properties and public properties.
The former was a superset of the latter and contained private
information that could end up in, for example, the audit log. This is
not needed for Cody's telemetry. See
https://sourcegraph.slack.com/archives/CN4FC7XT4/p1690356187459289 for
more context.
- Also removes unnecessary passing along of the serverEndpoint (which is
added globally).
- Simplify LocalStorage.{get,set}AnonymousUserID.
- Remove the `setAnonymousUserID` call from MessageProvider because
there was no call to `getAnonymousUserID` that would depend on it being
called there.
- Combine into a single `LocalStorage.anonymousUserID` method and
document it.

## Test plan

Run the extension with `cody.debug.enable` on and verify in the output
channel that the logged events look correct.