remove experimental file-path hallucination detector
This was the experimental feature that highlighted nonexistent file paths in responses from the LLM. It was cool, but it adds complexity, bugs (around parsing and re-escaping Markdown), and a ripgrep dependency. We will explore similar features in the future.

See https://sourcegraph.slack.com/archives/C05AGQYD528/p1690416892404469?thread_ts=1690416734.784219&cid=C05AGQYD528 for context.