fix default of triggerMoreEagerly in code
The default of triggerMoreEagerly was changed in vscode/package.json but not in code. The prior change was sufficient to make the default true, but the code was misleading. (We could improve this by making the TypeScript code in configuration.ts be aware of the JSON Schema of our configuration.)