use the `rg` (ripgrep) built into VS Code (#406)
VS Code ships with an `rg` binary for the current platform. By using
this instead of bundling our own, we (1) reduce the extension bundle
size (which makes it faster for people to try out Cody in VS Code), and
(2) simplify and speed up our own builds.

VS Code may change the location of `rg`, but it is fairly unlikely
because it uses the `@vscode/ripgrep` package that is unlikely to
significantly change.

See the VS Code implementation of its builtin `rg` service at
https://sourcegraph.com/github.com/microsoft/vscode@28849849a9e28f2abcecb5fad62564942821a333/-/blob/src/vs/workbench/services/search/node/ripgrepTextSearchEngine.ts.



## Test plan

Run a keyword context operation (such as a chat in a workspace with no
embeddings) and confirm that it works and that it invokes the VS
Code-bundled `rg`.