Completion: fix bad trailing `}` completion  (#378)
Fixes bad completion case
https://github.com/sourcegraph/cody/discussions/358#discussioncomment-6531096

After hours of catching corner cases, I think we can't rely on
regexp-based `shouldIncludeClosingLine` logic here. Counting brackets
doesn't provide much value when we don't know about the structure of the
expression where we try to insert completions. This PR just tries to
handle the simple case with an empty completion line with a closed
bracket symbol and cuts off this completion.

@philipp-spiess I don't think I fully understand how this is possible,
but unit tests tell me that completion should be cut off, but vs code
manual test shows that this problem is still there. Let me know if you
have a minute for a quick pair on this.

## Test plan
- Try to force `}` completion (see original GitHub discussion thread for
the example)
- Check that there is no `}` completion