disable telemetry when extension is running in dev or test mode
I noticed this interfering with my own "organic" telemetry and making it harder for me to understand my own usage of Cody. Now, only usage in the main VS Code window (not in a test, and not in a debug extension host instance) counts. This could increase the likelihood of errors in our actual telemetry, but that is mitigated by our heavy use of the Cody insiders build.