Autocomplete: Overhaul telemetry and fix various corner cases (#412)
Closes #178

This PR fixes a bunch of issues with the completions cache that should
make it much more easy to understand now (although there are a few
corner cases). The analytics code is quite tangled with the
implementation right now but I want to get the logic right first before
we do some cleanup.

Conceptually, the big thing that has changed is that we now **always
create a new completion entry** for new completion requests. **The only
exception to that is if the cache would resolve to the very same
completion that was already displayed**. This special case is necessary
to not send too many completion requests if we forward-type a completion
like so:


https://github.com/sourcegraph/cody/assets/458591/74f5c652-b7d8-49a3-896f-4e731234e55c

There are a bunch of other small fixes that I noticed were necessary
that I’m going to explain inline.

## Test plan

- Open the Output tabs and start playing with completions while keeping
track of how the console looks.
- We'll have to see how big the difference is in our metrics after
pushing this to prod.


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->