Autocomplete: Fix loading indicator and deletion opt-out (#456)
Two quick fixes for regressions I introduced in #412

- Loading indicator felt flaky so this this PR makes sure we start the
indicator a bit earlier and terminate in any case.
- The opt-out for a deletion case was wrong, causing it to skip the
condition if the last action was a deletion and there was no cache
response. This was not intended to be changed for now. The correct
behavior should still be: If there is a deletion, ONLY continue if twe
have an exact cache response. Otherwise return no completions.

## Test plan

- tested locally,

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->