move provider to vscodeInlineCompletionItemProvider.ts (#470)
No impl changes. This is just so that we can better separate the VS
Code-specific stuff from the rest (and helps with future commits).

This is an intermediate commit that will make our history cleaner. I
submitted it as a separate PR so that the Git history includes the
rename as an atomic operation.