never build deasync to speed up and de-flake builds
The completions-review-tool uses remix, which uses deasync, which only has prebuilt binaries for older Node.js versions (https://github.com/abbr/deasync-bin). This means that CI takes 1-2 extra minutes and is flakier due to node-gyp being invoked each time. This commit skips the node-gyp build of deasync. It may break some aspects of the completion-review-tool (although it seems to work fine), but that's an acceptable tradeoff for now because that is a dev-only tool and we can revert this commit if it is needed.