try to show completion when suggest widget is showing (#478)
Even when
`cody.autocomplete.experimental.completeSuggestWidgetSelection` is
false, still try to return a completion even if the suggest widget is
showing. This completion will often not be shown to the user (but we now
check for that case in telemetry). This removes complexity from the
provider code.

## Test plan

Set `editor.inlineSuggest.suppressSuggestions` to true. See that Cody
autocomplete shows up when the insertText of the suggest widget
selection matches the prefix of the completion (even without enabling
the Cody experimental setting).