return completions from the start of current line to reduce VS Code UI jitter (#482)
This reduces UI jitter in VS Code; when typing or deleting individual
characters, VS Code reuses the existing completion while it waits for
the new one to come in. This is also accomplished by our existing
caching when forward-typing, but this even further reduces it.



## Test plan

Type `const country =` and wait for it to suggest something like
`'USA'`. Type the first few letters of the suggestion then delete. With
this change, the suggestion still displays. Without this change, it
disappears.