Autocomplete: Remove trailing whitespace (#542)
I noticed some odd duplications in multi-line completion snippets that
was caused by a difference only in trailing whitespace. Since my IDE
does this on save automatically anyways, I figure it's save to filter
the LLM output for this. I noticed this to be an issue in both Anthropic
and StarCoder, but Anthropic encounters it more often.

## Test plan

- Added a test case


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->