fix problem with github copilot suppressing ctrl+enter in comment UI
In the VS Code comment UI, ctrl+enter/cmd+enter should submit the comment. Unfortunately, when GitHub Copilot is activated, it does nothing. This is likely a bug in GitHub Copilot's overly greedy keyboard bindings (it should have `&& !commentEditorFocused` in the `when` for `github.copilot.generate`).

To make that keybinding submit the comment in VS Code's comment UI again, work around this by disabling that keybinding when Cody is running (it is not possible to be more selective about when we disable the keybinding). This is reasonable because that is not a frequently used keybinding anymore for Copilot.