remove obsolete "Fixup Code from Inline Instructions" (#549)
This was the feature that let you write natural language interspersed in
your code, such as:

```
people.sorted().map(get first names)
```

and Cody would write the code, guessing your intent. This is several
months old, and it was very experimental. We've since moved to a much
more solid and prescriptive (in a good way) direction for inline chat
and fixups. This recipe is obsolete and misleading now, so it should be
removed.



## Test plan

Ensure that the `Fixup from Inline Instructions` recipe does not show up
anywhere. Test that `/fix` in inline chat still works.