fix ~/.vscode/cody.json recipes file watch that exhausted system
This fixes an issue where Cody (on Linux at least) would watch the user's home directory and exhaust the system's open files limit. This appears to be due to a bug or limitation in VS Code's `createFileSystemWatcher` API where `RelativePattern` globs are not optimized and instead set up recursive watchers on the first arg given (which, for the user's home dir, causes a very large number to be created).

To reproduce the issue, uncomment this code and then look for the following in the VS Code logs for file watchers (see below for how to enable these):

```
TRACE [File Watcher (parcel)] Request to start watching: /home/sqs
```

(`/home/sqs` is my home directory, and the file watcher then starts to traverse all dirs and set up ~250,000+ inotify watches.)

See https://github.com/microsoft/vscode/wiki/File-Watcher-Issues#logging-local for instructions on how to enable logging of file watchers to reproduce this.