Agent: add support for autocomplete
This PR adds a new `autocomplete/execute` endpoint to the agent
protocol. To implement this endpoint, the agent now directly depend on
the VS Code extension by providing a shim for all imports against the
`'vscode'` module. By using a shim, the agent can launch the VS Code
extension in a headless environment and capture the instance o the
completion provider when it's registered.

This PR doesn't implement all of VS Code APIs yet. It only implements
that necessary parts to get autocomplete working.