Autocomplete: Bring back completion synthesization from prior requests and retest all inflight requests (#559)
This PR brings back a completion cache that is able to synthesize
results even if the request has changed. It's much simpler than the old
request cache and only keeps a maximum of 50 completions around.

When the cache is queried, we compare the document position with all
cached completions for the same file and if there's something
salvageable, we synthesize a new completion base off of the previous
one.

The data structures are kept much simpler now. Instead of having a map
of multiple arrays of completions, we flattened the data structure to
only be an array of completions. Inflight request logging was also
changed from a per-file map to a flat set.

## Todo

I noticed what appears to be a regression that I’m still investigating.
Sometimes the last candidate response will _not_ be the actual last
completion that was visible. You can notice this happening at ~26sec in
this video:



https://github.com/sourcegraph/cody/assets/458591/8bea416e-28a9-4d67-8617-74ea19c802f8



## Test plan

Type a `console.log` statement, leafing only a _tiny_ pause after one
character. Observe that the response is going to be marked as
`CacheAfterRequestStarted`:

<img width="1610" alt="Screenshot 2023-08-03 at 14 14 14"
src="https://github.com/sourcegraph/cody/assets/458591/7534926a-bffb-491c-aaa8-69a74c4a837d">

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->