support signing in on VS Code Web (#625)
Refactor vscode input boxes for instance URL and access token.

Refactor the code to make it possible to just show an access token input
box.

Do not show Cody app login when running in VS Code Web

Support VS Code Web in login flow: VS Code Web does not support
callbacks like vscode:// from our access token creation flow because it
is running on the web and does not have a system protocol handler
registered. There is a solution to this, but it requires changes on the
Sourcegraph.com backend. So, for now, just open up the new access token
creation page and ask the user for the token.


## Test plan

Start with a fresh VS Code instance and try signing in via
Sourcegraph.com or via a token.