Autocomplete: Fix suggestion event over counting (#649)
This PR fixes an issue that sometimes caused completion suggestion
events to not fire correctly. The issue here was that we decided wether
or not a completion is visible in the `getInlineCompletions` code that
was now change to _still yield completions_.

The yielding is fine, we have relied on these results to be used for the
last candidate cache. However, this would also mean that we would store
a log id that was _deemed as not visible_ but that never reconsidered
this condition. This meant that such a completion could become visible
later and we would be able to accept it, even though it was never logged
as suggested.

To fix this, we move this logic into the VS Code specific part (which
conceptually makes more sense anyways since it's full of VS Code
specific logic) and also test the visibility after every cache
retrieval.

## Test plan

I've added three conditions to test invariants that should not be
possible. To test locally, I've added debugger break points into either
of them and repeatedly triggered a bazillion of completions. After these
changes, I don't seem to be able to trigger them anymore. I’m leaving
some logs in there, though, so we can see if they happen on production.

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->