JetBrains: add autocomplete support via Agent (#55638)
* JetBrains: add autocomplete support via Agent

Previously, we unconditionally used the custom autocomplete
implementation regardless if the agent was enabled or not. This commit
adds support to use agent-powered autocomplete when the agent is
enabled.

* Correctly process autocomplete results from the agent

Previously, the IntelliJ plugin had to post-process the LLM completions
resulting in custom logic for how we applied completions. This commit
adds a separate code path for interpreting agent completions (how we
render inlay hints) as well as how we accept agent completions (we
use `range` and `insertText` directly)

* IntelliJ extension send configuration to agent for experimental providers (#55776)

wip

* Skip completions with change deltas

* Reformat

* Fix formatting

* Remove unnecessary dependency

* Minor fixes in preparation to merge the PR

- Use `cody.autocomplete` setting names
- Log accepted agent autocomplete

* Remove unnecessary code and document some methods

* Reformat

* Mark autocomplete as displayed right before displaying

* Don't log event when displaying

---------

Co-authored-by: Chris Warwick <christopher.warwick@sourcegraph.com>
Former-commit-id: f79c0f8a96315b1b59861a460b2d89886a38b5a1