overhaul autocomplete provider config (#650)
- Show error messages during provider creation in the output channel
instead of (a) in a notification (which is annoying and shows up for
each keystroke) or (b) in the webview, which is not somewhere the user
would necessarily look for autocomplete-related issues.
- Do not default to the Anthropic provider if an invalid provider is
specified. If the user is specifying a provider in
`cody.autocomplete.advanced.provider`, their intent is likely to use a
non-default provider, and falling back to Anthropic would not be
desired.
- Remove the duplication in calls to `createCompletionsProvider`.
- Update settings docstrings.

(My initial motivation for this is that I got annoying VS Code
notifications while hacking on a new unstable autocomplete provider
based on Ollama.)

## Test plan

Set `cody.autocomplete.advanced.provider` to non-default and invalid
values and ensure that the configuration is auto-applied and that the
correct error messages are shown in the output channel.