Agent: add support for autocomplete (#624)
This PR adds a new `autocomplete/execute` endpoint to the agent
protocol. To implement this endpoint, the agent now directly depend on
the VS Code extension by providing a shim for all imports against the
`'vscode'` module. By using a shim, the agent can launch the VS Code
extension in a headless environment and capture the instance o the
completion provider when it's registered.

This PR doesn't implement all of VS Code APIs yet. It only implements
that necessary parts to get autocomplete working.


## Test plan

See new `index.test.ts`
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>
Co-authored-by: Chris Warwick <christopher.warwick@sourcegraph.com>
Co-authored-by: Stephen Gutekanst <stephen.gutekanst@gmail.com>