Agent: code cleanup and get autocomplete working again (#669)
Previously, autocomplete via agent wasn't working because of regressions
that got introduced in the last PR to fix ESLint errors. We have
integration tests that caught this regression but we don't run that test
in CI because it currently requires SRC_ACCESS_TOKEN.

This PR fixes the regression above along with other improvements:

- Use latest Node.js for agent binaries because the VSC extension is
using APIs that are not present in Node.js v16 (version we previously
used for the agent) causing the agent to crash.
- Remove leftover `console.log`
- Fix failing integration test after we added a new agent recipe
- Fail fast when running the extension via the agent and we don't
register the completion provider. We currently await on the completion
provider, which silently times out if we don't register the completion.


## Test plan

Manually tested with
https://github.com/sourcegraph/sourcegraph/pull/55826

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->