JetBrains: fix path of agent binaries in plugin zip (#55867)
The 3.0.9 was a failure because the agent binaries were located in the
`agent/agent/` directory instead of `agent/`. This code path was working
correctly at one point but it was only manually tested. This commit
fixes the issue and adds new assertions to the release script to prevent
a regression like this in the future.

Former-commit-id: 3fa2a0007b642459d976b0fe4ce527ffb8a26088