Don't require Esc to dismiss Cody menu (#700)
This removes the `ignoreFocusOut: true` settings on our command
quickpick, removing the need for you use to press `esc` in the input to
dismiss it. This makes it feel just as quick and light as the standard
`cmd-shift-p` command palette.

The one downside to removing it is if you need to change your code
selection mid-quickpick-item-selection, but given the standard command
palette has commands that depend on code selection too (e.g. "Transform
to Kebab Case") I think we should follow their suit and not introduce
the extra bit of friction that requiring `esc` adds.

## Test plan

- Opened command palette
- Clicked outside it and verified it dismisses