Use CSS scroll anchoring for most chat scroll pinning (#704)
Chat scroll pinning was broken for `/search` because that adds messages
in quick succession and the React effect was only looking at the change
in message author.

The general approach was also iffy because we were basically polling the
scroll position on updates, but in practice scrolling works best when it
is managed asynchronously by the browser.

This takes a different approach:
- Use CSS scroll anchoring. This takes care of 90% of what we want with
pinning the scroll position to the bottom: If you scroll away from the
anchor, the scroll position doesn't change, the browser manages all of
this for us. We also get better behavior when you resize the panel; the
scroll position will stick to the bottom pretty intelligently.
- Because browsers anchor the scroll position to the top until there's
any scrolling, this uses an intersection observer to detect when the
content becomes scrollable and force it to scroll. I actually like the
behavior without that, but have implemented this for consistency with
what we had.
- When there's a new human message, scroll to the bottom. This means
when you commit input in the chat box, we scroll to it.

I've made some subtle UX changes like animating some of the scrolls.

The scroll anchor adds a 5px gutter at the bottom of the content which
is the "grippy" region for the scroll. We can avoid that extra space if
we want, I took a look at it, it requires a couple of stacking contexts
and I thought it wasn't worth it. But we can do it if we want.

## Test plan

1. New Cody chat, ask it to enumerate 50 states in order of accession in
a numbered list. The scroll should follow the content as it streams in.
2. Scroll up, chat to ask it to enumerate the 50 states and their state
vegetables. The scroll should jump to the new question when you hit
enter; then follow the content.
3. Scroll up, switch to the files panel, switch back. The scroll
position should be where you left it.
4. While content is streaming in, scroll away from it. As new content
appears the scroll should not jump back to the bottom.
5. Scroll to the bottom, ask to enumerate 50 states etc., while content
is streaming in switch to the files panel, wait, then switch back. The
scroll position should be where it was when you switched away. That is,
*not* pinned any more.
6. Do `/search foo` and the scroll position should keep up with the
bottom of the content. (This comes in very *fast* and I think we should
add a heuristic to maybe not do that. But it is better than the behavior
today, which quietly drops the new content below the fold.)
7. Reload the webviews/restart the extension with a long transcript. The
restored transcript should be scrolled to the bottom.