Autocomplete: Improve interaction with completions menu (#636)
This PR extends the existing
`cody.autocomplete.experimental.completeSuggestWidgetSelection` to take
full advantage of the suggest widget (in the best way that seems to be
possible) and use the selected item to guide autocomplete requests.

There is one instance when this does not work (which is accounted for in
the code) that happens when the current code in the editor does not
overlap with the current selection already. Consider this case:

<img width="748" alt="Screenshot 2023-08-14 at 15 45 51"
src="https://github.com/sourcegraph/cody/assets/458591/54901557-64c2-486c-83da-88b60c078f6d">

In this case, the current text content contain `console.l` but the
suggestion is pointed to an item that would require the prefix to start
with `console.d` instead. _In this case, VS Code will never render a
ghost text._ Because of that, I've added a top-line bail for this
specific case.

## Test plan

Enable `cody.autocomplete.experimental.completeSuggestWidgetSelection`:


https://github.com/sourcegraph/cody/assets/458591/eaf73d66-4016-4251-8bff-d8410b01ddac



<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->