Agent: add requests for logging events and to find user ID (#734)
Previously, the JetBrains plugin still sent direct network requests to
the Sourcegraph instance. This was problematic because it meant that
users behind proxies had to configure both proxy settings for the agent
and for the Java plugin. This PR adds two endpoints to the Agent
protocol allowing the JetBrains plugin to send requests to the Agent
instead of sending network requests to Sourcegraph.

## Test plan

Tested alongside JetBrains PR
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->