Agent: disable telemetry only via the shim
Previously, we used the `isRunningInsideAgent` config to determine
whether telemetry should be disable or not. This solution had the
problem that we actually want to send telemetry from the agent when the
client sends the recently added `graphql/logEvent` request. We just want
to disable telemetry for vscode-specific events.

This PR fixes the problem by reverting the previous
`isRunningInsideAgent` check and adds a check based on the
`cody.telemetry.level` configuration setting. The shim hardcodes the
value to `'off'` making telemetry disabled when running inside the
agent, while the telmetry level defaults to `'all'` when we manually
construct the GraphQL client.