Move typewriter so it does not throttle keyword expansion, reranker (#719)
The Typewriter was integrated at a convenient point for other agents
possibly, but it didn't work well for Cody-VScode. VScode's
BotResponseMultiplexer buffers content to parse tags in the output,
which chunks the incremental effect a bit. The typewriter also slowed
down consumers like inline fixups which just want the complete code and
don't care if it appears "typed."

Since then, Cody-VScode has added LLM turns for keyword expansion and
context re-ranking. The typewriter effect also applied to these,
needlessly slowing them down. For me these turns are about twice as fast
now the typewriter is out of their path.

This moves the typewriter up to the MessageProvider, and only applies
the typewriter effect to output destined for the chat sidebar. This
makes all the code consumers faster because there's no animation delay.
(The "typing" is very smooth now—maybe too smooth—because there's no
coincidental chunking from the BotResponseMultiplexer.)

This adds a way for the consumer to drain the buffered content out of
the typewriter while still "animating" it, `flush`. It indicates how
much backpressure there was on the typewriter that we didn't really
notice this "chunk at the end" behavior.

See #714.

## Test plan

```
pnpm test:unit && pnpm test:integration && pnpm test:e2e
```

Manual test:

1. Open Cody chat sidebar.
2. Ask it to enumerate the 50 states of the US in order of accession
with their state flowers in a numbered list

The output should appear incrementally and completely.

---------

Co-authored-by: Naman Kumar <naman@outlook.in>