Autocomplete: Abort network requests when a completion request is resolved with a previously started request's response (#762)
Right now, a completion can be resolved while the network request is
still being in progress if a previous completion request matches the new
editor text. This happens in ~10% of situations and was responsible for
a small increase in latency.

I noticed that we did not abort the request if that is the case though
and we still used the second response to the seed the cache. I think in
this case, this is not necessary since we already have a completion that
matches and that we consider a "match" for the given prefix. I think
that we should rather preserve the costs and network bandwidth in this
case.

To think through this:

- 👉 User starts by typing `console.l` and waits a few MS
- ➡️ We start request 1 with the prefix `console.`
- 👉 User continues to type until it reaches `console.log`
- ➡️ We start request 2 with the prefix `console.log`
- ⬅️ Request 1 resolves with `console.log('hello world')`
- We detect this can also apply to request 2 so we resolve request
number 1 and 2 with this value
- ⬅️ (later) Request 2 resolves with something else (e.g.
`console.log('foobar')`). Since 2 was already resolved, we only put this
into the network request for the exact `console.log` prefix that we
already resolved before.

## Test plan

Added a test case to ensure the abort signal is being propagated in that
case

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->