fix: support /fix command in chat (#790)
fix: handle /fix command by executing `cody.fixup.new`

This PR includes:
- The `/fix` command now executes the `cody.fixup.new` command during
the `chatCommandFilter` instead of executing the fixup recipe that
returns an error (see image below). This allows the fixup flow to be
handled in the correct way.
- The `cody.fixup.new` command now accepts the correct args to generate
a fixup task. Previously, it was only accepting `range`, which is used
as `options` and gets passed down incorrectly.


## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

1. Start Cody from this branch
2. Type `/fix` with instruction (e.g. /fix rename function to foo) in
sidebar chat

###. Before: An error pop up saying "select some code to fixup" even
when you have an active selection


![image](https://github.com/sourcegraph/cody/assets/68532117/d3b1198f-6f21-411e-b98e-392ceb55e822)

### After: fixup gets executed correctly


![image](https://github.com/sourcegraph/cody/assets/68532117/8501494b-133c-4156-ac93-b4ee678c894e)